+++ server/manage_db_tools/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ManageDbToolsConfig(AppConfig):
+    name = 'manage_db_tools'