+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.6.0a Code refactoring
+
+- Added CSV download functionality
+- Added new app for tools
+- `authed.html` refactiring
+- Some functionality from home moved to tools app
+- Added `404.html` template
+
+
 ## 0.5.0a Full integration
 
 - Optimized integration
@@ -9,7 +18,7 @@
 
 ## 0.4.0a Added basic nn analyser algorithm
 
-- Add prototype of nn algorithm 
+- Add prototype of nn algorithm
 - Fix some bug with server
 
 
@@ -30,5 +39,5 @@
 
 - Added markdown pages support. Added privacy and api pages.
 - Added config.json file in root directory
-- Added CHANGELOG.md file for changelogs
-- Added analysis folser for files with nn 
+- Added `CHANGELOG.md` file for changelogs
+- Added analysis folser for files with nn